requirements.txt
fastapi>=0.101.0
uvicorn[standard]>=0.23.0
sqlalchemy>=2.0.0
alembic>=1.12.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
pandas>=2.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
xgboost>=2.0.0
lightgbm>=4.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
python-multipart>=0.0.6
bcrypt>=4.0.1
websockets